Module Http_multipart_formdata.Make

Types

Parameters

module P : Reparse.PARSER

Signature

type input = P.input

and 'a promise = 'a P.promise

and reader

and read_result = [

| `End

| `Header of part_header

| `Body of Cstruct.t

| `Body_end

| `Error of string

]

and part_header

Represents a parsed multipart part header data.

Multipart Reader

val reader : ?read_body_len:int -> boundary -> input -> reader

reader ?read_body_len boundary input creates reader. The default value for read_body_len is 1KB.

val read_part : reader -> read_result promise

read_part ?read_body_len ~boundary reader reads a http multipart body and returns a read_result.

Part header

val name : part_header -> string

name t returns the form field name

val content_type : part_header -> string

content_type t returns the part content-type.

val filename : part_header -> string option

filename t returns the uploaded filename is the multipart is a file

val header_value : string -> part_header -> string option

param_value name t returns the multipart parameter value with name name.

Pretty Printers

val pp_part_header : Stdlib.Format.formatter -> part_header -> unit

val pp_read_result : Stdlib.Format.formatter -> read_result -> unit