Module Http_multipart_formdata.Part

Represents a parsed multipart part. A part corresponds to a submitted form field data in a HTTP request.

type t = {

body : bytes;	(*
Body content
*)

name : string;	(*
Name of the part - form field name
*)

content_type : string;	(*
HTTP content type of the part body. "text/plain" is default
*)

filename : string option;	(*
filename form field attribute.
*)

parameters : string Map.t;	(*
Additional key = value params of the form field.
*)

}

val pp : Stdlib.Format.formatter -> t -> unit

pp fmt part is the pretty printer for t.

val equal : t -> t -> bool

equal part1 part2 returns true if part1 and part2 are equal.