Module Http_multipart_formdata.Part_header

Represents a parsed multipart part header data.

type t

val name : t -> string

name t returns the form field name

val content_type : t -> string

content_type t returns the part content-type.

val filename : t -> string option

filename t returns the uploaded filename is the multipart is a file

val param_value : string -> t -> string option

param_value name t returns the multipart parameter value with name name.

val compare : t -> t -> int

val equal : t -> t -> bool

val pp : Stdlib.Format.formatter -> t -> unit