Module Http_multipart_formdata

Part header

type part_header

Represents a parsed multipart part header data.

val name : part_header -> string

name t returns the form field name

val content_type : part_header -> string

content_type t returns the part content-type.

val filename : part_header -> string option

filename t returns the uploaded filename is the multipart is a file

val param_value : string -> part_header -> string option

param_value name t returns the multipart parameter value with name name.

val pp_part_header : Stdlib.Format.formatter -> part_header -> unit

Mulipart Boundary parser

type boundary = string

Represents the multipart boundary value.

val parse_boundary : content_type:string -> (boundary, string) Stdlib.result

parse_boundary ~content_type parses content_type to extract boundary value.content_type is the HTTP request Content-Type header value.

Multipart Parser

module type MULTIPART_PARSER = sig ... end

module Make (P : Reparse.PARSER) : MULTIPART_PARSER with type input = P.input with type 'a t = 'a P.t with type 'a promise = 'a P.promise