Module Http_multipart_formdata.Make

Parameters

module P : Reparse.PARSER

Signature

type input = P.input

type 'a t = 'a P.t

type 'a promise = 'a P.promise

type reader

and read_result = [

| `End

| `Header of part_header

| `Body of Cstruct.t

| `Body_end

| `Error of string

]

val reader : ?read_body_len:int -> boundary -> input -> reader

reader ?read_body_len boundary input creates reader. The default value for read_body_len is 1KB.

val parse_part : reader -> read_result promise

parse_part ?read_body_len ~boundary reader parse http multipart body and returns a read_result.

read_body_len determines the size of the multipart body to read in bytes. By default 1KB.

val pp_read_result : Stdlib.Format.formatter -> read_result -> unit